+++ CHANGELOG.md
@@ -1,5 +1,16 @@
 # Changelog
 
+## 0.23.0
+
+### Misc
+
+- Extended message statuses "wait", "sent", "read" and "lost"
+- Internal chat connection manager
+
+### Features
+
+- Implemented `chatStatus` state with "disconnected", "connecting" and "connected"
+- Implemented `processOnMessageSent` function to handle actions triggered by lost/sent message
 
 ## 0.22.0